+++ config/production.ts
@@ -4,7 +4,7 @@ import {readValidators} from './utils';
 const production: GlobalConfig = {
   ethereum: {
     contractRegistry: {
-      address: '0x8d16D5D2859f4c54b226180A46F26D57A4d727A0'
+      address: '0x968A798Be3F73228c66De06f7D1109D8790FB64D'
     },
     token: {
       totalSupply: '1000000' + '0'.repeat(18),
@@ -12,13 +12,13 @@ const production: GlobalConfig = {
       symbol: 'UMB'
     },
     chain: {
-      blockPadding: 6
+      padding: 60
     },
     validators: readValidators('https://validator.umb.network'),
   },
   smartchain: {
     contractRegistry: {
-      address: '0x67D0B6c70B71d1065927401744f853f60d6f59a5'
+      address: '0xa3A8D57DFeC72fe69FB0D7D1991e7B242De3603A'
     },
     token: {
       totalSupply: '1000000' + '0'.repeat(18),
@@ -26,7 +26,7 @@ const production: GlobalConfig = {
       symbol: 'UMB'
     },
     chain: {
-      blockPadding: 6
+      padding: 60
     },
     validators: readValidators('https://validator-bsc.umb.network'),
   }