+++ contracts/Chain.sol
@@ -3,54 +3,46 @@ pragma solidity ^0.6.8;
 pragma experimental ABIEncoderV2;
 
 import "@openzeppelin/contracts/math/SafeMath.sol";
-import "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
 import "@openzeppelin/contracts/cryptography/MerkleProof.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
+import "@umb-network/toolbox/dist/contracts/lib/ValueDecoder.sol";
 
-import "./lib/LeafDecoder.sol";
 import "./interfaces/IStakingBank.sol";
 import "./interfaces/IValidatorRegistry.sol";
 
 import "./extensions/Registrable.sol";
 import "./Registry.sol";
 
-contract Chain is ReentrancyGuard, Registrable, Ownable {
+contract Chain is Registrable, Ownable {
   using SafeMath for uint256;
-  using LeafDecoder for bytes;
+  using ValueDecoder for bytes;
 
   // ========== STATE VARIABLES ========== //
-
-  uint256 public blockPadding;
 
   bytes constant public ETH_PREFIX = "\x19Ethereum Signed Message:\n32";
 
   struct Block {
     bytes32 root;
-    address minter;
-    uint256 staked;
-    uint256 power;
-    uint256 anchor;
-    uint256 timestamp;
-    uint256 dataTimestamp;
+    uint32 dataTimestamp;
+    uint128 affidavit;
   }
 
-  struct ExtendedBlock {
-    Block data;
-    address[] voters;
-    mapping(address => uint256) votes;
-    mapping(bytes32 => uint256) numericFCD;
+  struct FirstClassData {
+    uint224 value;
+    uint32 dataTimestamp;
   }
 
-  mapping(uint256 => ExtendedBlock) public blocks;
+  mapping(uint256 => Block) public blocks;
+  mapping(bytes32 => FirstClassData) public fcds;
 
-  uint256 public blocksCount;
-  uint256 public blocksCountOffset;
+  uint32 public blocksCount;
+  uint32 public blocksCountOffset;
+  uint16 public padding;
 
   // ========== CONSTRUCTOR ========== //
 
-  constructor(address _contractRegistry, uint256 _blockPadding) public Registrable(_contractRegistry) {
-    blockPadding = _blockPadding;
-
+  constructor(address _contractRegistry, uint16 _padding) public Registrable(_contractRegistry) {
+    padding = _padding;
     Chain oldChain = Chain(Registry(_contractRegistry).getAddress("Chain"));
 
     if (address(oldChain) != address(0x0)) {
@@ -61,82 +53,62 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
 
   // ========== MUTATIVE FUNCTIONS ========== //
 
-  function setBlockPadding(uint256 _blockPadding) external onlyOwner {
-    blockPadding = _blockPadding;
-    emit LogBlockPadding(msg.sender, _blockPadding);
+  function setPadding(uint16 _padding) external onlyOwner {
+    padding = _padding;
+    emit LogPadding(msg.sender, _padding);
   }
 
   function submit(
-    uint256 _dataTimestamp,
+    uint32 _dataTimestamp,
     bytes32 _root,
     bytes32[] memory _keys,
     uint256[] memory _values,
     uint8[] memory _v,
     bytes32[] memory _r,
     bytes32[] memory _s
-  ) public nonReentrant returns (bool) {
-    uint256 blockHeight = getBlockHeight();
-    require(blocks[blockHeight].data.anchor == 0, "block already mined for current blockHeight");
-    // in future we can add timePadding and remove blockPadding
-    require(blocks[blockHeight - 1].data.dataTimestamp < _dataTimestamp, "can NOT submit older data");
-
-    bytes memory testimony = abi.encodePacked(blockHeight, _dataTimestamp, _root);
-
+  ) public {
+    uint32 lastBlockId = getLatestBlockId();
+    require(blocks[lastBlockId].dataTimestamp + padding < block.timestamp, "do not spam");
+    require(blocks[lastBlockId].dataTimestamp < _dataTimestamp, "can NOT submit older data");
+    require(_dataTimestamp <= block.timestamp, "oh, so you can predict future");
     require(_keys.length == _values.length, "numbers of keys and values not the same");
 
+    bytes memory testimony = abi.encodePacked(_dataTimestamp, _root);
+
     for (uint256 i = 0; i < _keys.length; i++) {
-      blocks[blockHeight].numericFCD[_keys[i]] = _values[i];
+      require(uint224(_values[i]) == _values[i], "FCD overflow");
+      fcds[_keys[i]] = FirstClassData(uint224(_values[i]), _dataTimestamp);
       testimony = abi.encodePacked(testimony, _keys[i], _values[i]);
     }
 
     bytes32 affidavit = keccak256(testimony);
+    uint256 power = 0;
 
-    (blocks[blockHeight].data.staked, blocks[blockHeight].data.power) =
-      _validateSignatures(blockHeight, affidavit, _v, _r, _s);
-
-    blocks[blockHeight].data.root = _root;
-    blocks[blockHeight].data.minter = msg.sender; //TODO check if validator is registered
-    blocks[blockHeight].data.anchor = block.number;
-    blocks[blockHeight].data.timestamp = block.timestamp;
-    blocks[blockHeight].data.dataTimestamp = _dataTimestamp;
-
-    blocksCount++;
-
-    emit LogMint(msg.sender, blockHeight, block.number);
-
-    return true;
-  }
-
-  function _validateSignatures(
-    uint256 _blockHeight,
-    bytes32 _affidavit,
-    uint8[] memory _v,
-    bytes32[] memory _r,
-    bytes32[] memory _s
-  ) internal returns (uint256 staked, uint256 power){
     IStakingBank stakingBank = stakingBankContract();
-    staked = stakingBank.totalSupply();
-
-    power = 0;
+    uint256 staked = stakingBank.totalSupply();
+    address prevSigner = address(0x0);
 
     for (uint256 i = 0; i < _v.length; i++) {
-      address signer = recoverSigner(_affidavit, _v[i], _r[i], _s[i]);
+      address signer = recoverSigner(affidavit, _v[i], _r[i], _s[i]);
       uint256 balance = stakingBank.balanceOf(signer);
-      require(blocks[_blockHeight].votes[signer] == 0, "validator included more than once");
 
-      if (balance == 0) {
-        // if no balance -> move on, it can be invalid address but also fresh validator with no balance
-        // we can spend gas and check if address is validator address, but I see no point, its cheaper to ignore
-        // if we calculated root for other blockHeight, then recovering signer will not work -> move on
-        continue;
-      }
+      require(prevSigner < signer, "validator included more than once");
+      prevSigner = signer;
+      if (balance == 0) continue;
 
-      blocks[_blockHeight].voters.push(signer);
-      blocks[_blockHeight].votes[signer] = balance;
+      emit LogVoter(lastBlockId + 1, signer, balance);
       power = power.add(balance);
     }
 
-    require(power.mul(100) > staked.mul(66), "not enough power was gathered");
+    // TODO to optimise, break the loop when get enough power
+    require(power.mul(100) >= staked.mul(66), "not enough power was gathered");
+
+    blocks[lastBlockId + 1].root = _root;
+    blocks[lastBlockId + 1].dataTimestamp = _dataTimestamp;
+    blocks[lastBlockId + 1].affidavit = uint128(bytes16(affidavit));
+    blocksCount++;
+
+    emit LogMint(msg.sender, lastBlockId + 1, staked, power);
   }
 
   // ========== VIEWS ========== //
@@ -145,25 +117,27 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
     return "Chain";
   }
 
-  function recoverSigner(bytes32 affidavit, uint8 _v, bytes32 _r, bytes32 _s) public pure returns (address) {
-    bytes32 hash = keccak256(abi.encodePacked(ETH_PREFIX, affidavit));
+  function recoverSigner(bytes32 _affidavit, uint8 _v, bytes32 _r, bytes32 _s) public pure returns (address) {
+    bytes32 hash = keccak256(abi.encodePacked(ETH_PREFIX, _affidavit));
     return ecrecover(hash, _v, _r, _s);
   }
 
   function getStatus() external view returns(
     uint256 blockNumber,
-    uint256 lastDataTimestamp,
-    uint256 lastBlockHeight,
+    uint16 timePadding,
+    uint32 lastDataTimestamp,
+    uint32 lastBlockId,
     address nextLeader,
-    uint256 nextBlockHeight,
+    uint32 nextBlockId,
     address[] memory validators,
     uint256[] memory powers,
     string[] memory locations,
     uint256 staked
   ) {
     blockNumber = block.number;
-    lastBlockHeight = getLatestBlockHeightWithData();
-    lastDataTimestamp = blocks[lastBlockHeight].data.dataTimestamp;
+    timePadding = padding;
+    lastBlockId = getLatestBlockId();
+    lastDataTimestamp = blocks[lastBlockId].dataTimestamp;
 
     IValidatorRegistry vr = validatorRegistryContract();
     uint256 numberOfValidators = vr.getNumberOfValidators();
@@ -175,7 +149,9 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
       (, locations[i]) = vr.validators(validators[i]);
     }
 
-    nextLeader = numberOfValidators > 0 ? validators[getLeaderIndex(numberOfValidators, blockNumber + 1)] : address(0);
+    nextLeader = numberOfValidators > 0
+      ? validators[getLeaderIndex(numberOfValidators, block.timestamp + 1)]
+      : address(0);
 
     IStakingBank stakingBank = stakingBankContract();
     powers = new uint256[](numberOfValidators);
@@ -185,49 +161,50 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
       powers[i] = stakingBank.balanceOf(validators[i]);
     }
 
-    nextBlockHeight = getBlockHeightForBlock(blockNumber + 1);
+    nextBlockId = getBlockId();
   }
 
-  function getBlockHeight() public view returns (uint256) {
-    return getBlockHeightForBlock(block.number);
-  }
-
-  function getBlockHeightForBlock(uint256 _ethBlockNumber) public view returns (uint256) {
-    uint _blocksCount = blocksCount + blocksCountOffset;
+  function getBlockId() public view returns (uint32) {
+    uint32 _blocksCount = blocksCount + blocksCountOffset;
 
     if (_blocksCount == 0) {
       return 0;
     }
 
-    return (blocks[_blocksCount - 1].data.anchor + blockPadding < _ethBlockNumber)
-      ? _blocksCount
-      : _blocksCount - 1;
+    if (blocks[_blocksCount - 1].dataTimestamp + padding < block.timestamp) {
+      return _blocksCount;
+    }
+
+    return _blocksCount - 1;
   }
 
-  function getLatestBlockHeightWithData() public view returns (uint256) {
+  function getLatestBlockId() public view returns (uint32) {
     return blocksCount + blocksCountOffset - 1;
   }
 
-  function getLeaderIndex(uint256 numberOfValidators, uint256 ethBlockNumber) public view returns (uint256) {
-    uint256 latestBlockHeight = getLatestBlockHeightWithData();
+  // note: I think its time to move leader selection from Chain
+  // we don't have anchor so we using timestamp, but timestamp if available without calling blockchain
+  // I will leave this methods but it should be deprecated soon
+  function getLeaderIndex(uint256 _numberOfValidators, uint256 _timestamp) public view returns (uint256) {
+    uint32 latestBlockId = getLatestBlockId();
 
-    // blockPadding + 1 => because padding is a space between blocks, so next round starts on first block after padding
-    uint256 validatorIndex = latestBlockHeight +
-      (ethBlockNumber - blocks[latestBlockHeight].data.anchor) / (blockPadding + 1);
+    // timePadding + 1 => because padding is a space between blocks, so next round starts on first block after padding
+    uint256 validatorIndex = latestBlockId +
+      (_timestamp - blocks[latestBlockId].dataTimestamp) / (padding + 1);
 
-    return validatorIndex % numberOfValidators;
+    return uint16(validatorIndex % _numberOfValidators);
   }
 
   function getNextLeaderAddress() public view returns (address) {
-    return getLeaderAddressAtBlock(block.number + 1);
+    return getLeaderAddressAtTime(block.timestamp + 1);
   }
 
   function getLeaderAddress() public view returns (address) {
-    return getLeaderAddressAtBlock(block.number);
+    return getLeaderAddressAtTime(block.timestamp);
   }
 
   // @todo - properly handled non-enabled validators, newly added validators, and validators with low stake
-  function getLeaderAddressAtBlock(uint256 ethBlockNumber) public view returns (address) {
+  function getLeaderAddressAtTime(uint256 _timestamp) public view returns (address) {
     IValidatorRegistry validatorRegistry = validatorRegistryContract();
 
     uint256 numberOfValidators = validatorRegistry.getNumberOfValidators();
@@ -236,12 +213,12 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
       return address(0x0);
     }
 
-    uint256 validatorIndex = getLeaderIndex(numberOfValidators, ethBlockNumber);
+    uint256 validatorIndex = getLeaderIndex(numberOfValidators, _timestamp);
 
     return validatorRegistry.addresses(validatorIndex);
   }
 
-  function verifyProof(bytes32[] memory _proof, bytes32 _root, bytes32 _leaf) public pure returns (bool) {
+  function verifyProof(bytes32[] memory _proof, bytes32 _root, bytes32 _leaf) public view returns (bool) {
     if (_root == bytes32(0)) {
       return false;
     }
@@ -249,17 +226,17 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
     return MerkleProof.verify(_proof, _root, _leaf);
   }
 
-  function hashLeaf(bytes memory _key, bytes memory _value) public pure returns (bytes32) {
+  function hashLeaf(bytes memory _key, bytes memory _value) public view returns (bytes32) {
     return keccak256(abi.encodePacked(_key, _value));
   }
 
   function verifyProofForBlock(
-    uint256 _blockHeight,
+    uint256 _blockId,
     bytes32[] memory _proof,
     bytes memory _key,
     bytes memory _value
   ) public view returns (bool) {
-    return verifyProof(_proof, blocks[_blockHeight].data.root, hashLeaf(_key, _value));
+    return verifyProof(_proof, blocks[_blockId].root, keccak256(abi.encodePacked(_key, _value)));
   }
 
   function bytesToBytes32Array(
@@ -284,7 +261,7 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
   }
 
   function verifyProofs(
-    uint256[] memory _blockHeights,
+    uint32[] memory _blockIds,
     bytes memory _proofs,
     uint256[] memory _proofItemsCounter,
     bytes32[] memory _leaves
@@ -295,7 +272,7 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
     for (uint256 i = 0; i < _leaves.length; i++) {
       results[i] = verifyProof(
         bytesToBytes32Array(_proofs, offset, _proofItemsCounter[i]),
-        blocks[_blockHeights[i]].data.root,
+        blocks[_blockIds[i]].root,
         _leaves[i]
       );
 
@@ -303,98 +280,34 @@ contract Chain is ReentrancyGuard, Registrable, Ownable {
     }
   }
 
-  function decodeLeafToNumber(bytes memory _leaf) public pure returns (uint) {
-    return _leaf.leafToUint();
+  function getBlockRoot(uint32 _blockId) external view returns (bytes32) {
+    return blocks[_blockId].root;
   }
 
-  function decodeLeafToFloat(bytes memory _leaf) public pure returns (uint) {
-    return _leaf.leafTo18DecimalsFloat();
+  function getBlockTimestamp(uint32 _blockId) external view returns (uint32) {
+    return blocks[_blockId].dataTimestamp;
   }
 
-  function verifyProofForBlockForNumber(
-    uint256 _blockHeight,
-    bytes32[] memory _proof,
-    bytes memory _key,
-    bytes memory _value
-  ) public view returns (bool, uint256) {
-    return (verifyProof(_proof, blocks[_blockHeight].data.root, hashLeaf(_key, _value)), _value.leafToUint());
-  }
-
-  function verifyProofForBlockForFloat(
-    uint256 _blockHeight,
-    bytes32[] memory _proof,
-    bytes memory _key,
-    bytes memory _value
-  ) public view returns (bool, uint256) {
-    return (
-      verifyProof(_proof, blocks[_blockHeight].data.root, hashLeaf(_key, _value)),
-      _value.leafTo18DecimalsFloat()
-    );
-  }
-
-  function getBlockData(uint256 _blockHeight) external view returns (Block memory) {
-    return blocks[_blockHeight].data;
-  }
-
-  function getBlockRoot(uint256 _blockHeight) external view returns (bytes32) {
-    return blocks[_blockHeight].data.root;
-  }
-
-  function getBlockMinter(uint256 _blockHeight) external view returns (address) {
-    return blocks[_blockHeight].data.minter;
-  }
-
-  function getBlockStaked(uint256 _blockHeight) external view returns (uint256) {
-    return blocks[_blockHeight].data.staked;
-  }
-
-  function getBlockPower(uint256 _blockHeight) external view returns (uint256) {
-    return blocks[_blockHeight].data.power;
-  }
-
-  function getBlockAnchor(uint256 _blockHeight) external view returns (uint256) {
-    return blocks[_blockHeight].data.anchor;
-  }
-
-  function getBlockTimestamp(uint256 _blockHeight) external view returns (uint256) {
-    return blocks[_blockHeight].data.timestamp;
-  }
-
-  function getBlockVotersCount(uint256 _blockHeight) external view returns (uint256) {
-    return blocks[_blockHeight].voters.length;
-  }
-
-  function getBlockVoters(uint256 _blockHeight) external view returns (address[] memory) {
-    return blocks[_blockHeight].voters;
-  }
-
-  function getBlockVotes(uint256 _blockHeight, address _voter) external view returns (uint256) {
-    return blocks[_blockHeight].votes[_voter];
-  }
-
-  function getNumericFCD(uint256 _blockHeight, bytes32 _key) public view returns (uint256 value, uint timestamp) {
-    ExtendedBlock storage extendedBlock = blocks[_blockHeight];
-    return (extendedBlock.numericFCD[_key], extendedBlock.data.timestamp);
-  }
-
-  function getNumericFCDs(
-    uint256 _blockHeight, bytes32[] calldata _keys
-  ) external view returns (uint256[] memory values, uint256 timestamp) {
-    timestamp = blocks[_blockHeight].data.timestamp;
+  function getCurrentValues(bytes32[] calldata _keys)
+  external view returns (uint256[] memory values, uint32[] memory timestamps) {
+    timestamps = new uint32[](_keys.length);
     values = new uint256[](_keys.length);
 
     for (uint i=0; i<_keys.length; i++) {
-      values[i] = blocks[_blockHeight].numericFCD[_keys[i]];
+      FirstClassData storage numericFCD = fcds[_keys[i]];
+      values[i] = uint256(numericFCD.value);
+      timestamps[i] = numericFCD.dataTimestamp;
     }
   }
 
-  function getCurrentValue(bytes32 _key) external view returns (uint256 value, uint timestamp) {
-    // it will revert when no blocks
-    return getNumericFCD(getLatestBlockHeightWithData(), _key);
+  function getCurrentValue(bytes32 _key) external view returns (uint256 value, uint256 timestamp) {
+    FirstClassData storage numericFCD = fcds[_key];
+    return (uint256(numericFCD.value), numericFCD.dataTimestamp);
   }
 
   // ========== EVENTS ========== //
 
-  event LogMint(address indexed minter, uint256 blockHeight, uint256 anchor);
-  event LogBlockPadding(address indexed executor, uint256 blockPadding);
+  event LogPadding(address indexed executor, uint16 timePadding);
+  event LogMint(address indexed minter, uint256 blockId, uint256 staked, uint256 power);
+  event LogVoter(uint256 indexed blockId, address indexed voter, uint256 vote);
 }