+++ scripts/deployRegistry.ts
@@ -1,10 +1,14 @@
+import { pressToContinue } from './utils/helpers';
+
 require('custom-env').env(); // eslint-disable-line
 
 import { deployContractRegistry } from './deployers/registry';
 
-deployContractRegistry()
-  .then(() => process.exit(0))
-  .catch((error) => {
-    console.error(error);
-    process.exit(1);
-  });
+pressToContinue('y', () => {
+  deployContractRegistry()
+    .then(() => process.exit(0))
+    .catch((error) => {
+      console.error(error);
+      process.exit(1);
+    });
+});