+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [2.0.0] - 2021-05-24
+
+### Changed
+- use `uint32` for `dataTimestamp`
+- remove FCD from block and save only last values
+- storage optimisation
+- use merkle tree library from SDK
+
+### Removed
+- remove `LeafDecoder`
+
 ## [1.0.3] - 2021-04-24
 ### Fixed
 - ensure `getLeaderIndex` uses padding in right way